Heavy refactor on types
Modeled the types by categorizing them into 3 kinds of types: Function types, Single types and Parametric types.
Function types can contain other types as its input and output.

By this model, all types are currified by default.
Most work is handled by the types themselves, with the applied method you can get the result of applying a single, parametric type or another function type to a function type (including replacing any parametric type in the original function type by concretes types provided)
This makes compositions and high order functions not be an special case.

The position is needed in several methods related with function types to allow applying parameters even when they are not the first one (since the model is currified by default this got a bit tricker and that parameter was needed).

Another change made was on how the checks are done on the blocks:
- Now there aren't different checks for functions, compositions and values. We only check the onChange for functions (compositions are functions so use this as well), and instead of checking the parent connections, we check the children connections (this covers the case of applying a value to a function).

The behaviour was changed a bit, now, if we are applying several parameters to a function, and we pass a parameter that doesn't type check, the one that will be bumped is the one we just tried to apply. The parameters that were already in the function are kept there. This was kind of accidental because I didn't know how to bump all of the parameters :P.