Fix bug related with parametric types
I detected an issue with this that we should fix. If we have a function with type variables in its type and we pass a value whose type also has type variables, if the type variables in both have the same names, we aren't performing any renaming so we could be losing variables.
For example,
```
Prelude> :t (id .)
(id .) :: (a -> c) -> a -> c
```

Would have resulted in (a -> a) -> (a -> a)

Which means that expressions like

`(id . even $ 0)` did not type check

The fix was done by modifying the parametric types that weren't replaced b concrete values when concretizing. So, what happens now when a parametric type is concretized is either: it gets replaced by a concrete value, or it earns a ' at the end of its type variable name.