Decouple our own types from Blockly Types
If we want to intersect and typecheck all applications ourselves, we'd have problems using Blockly types because some applications would be rejected by Blockly even before we can check the types ourselves. So if we wanted to return custom errors we woudln't be able to do so in the cases in which Blockly doesn't allow you to connect incompatible blocks.