Type checking by constraints generation + unification
Changed the type system so it delegates a lot of the work on types that were modeled.
Took the idea from this talk: https://www.youtube.com/watch?v=x3evzO8O9e8

The idea is, let's say we have something we want to typecheck like:

(even . id $ 0)

We want to know the composition's type (.) is correct in this case, how do we do that?

Let's check what's the type of the values involved:
(.) :: (b -> c) -> (a -> b) -> a -> c
id :: a -> a
even :: Number -> Bool
0 :: Number

A step 0 we do is taking all the functions involved and renaming the type variables names that are repeated in different functions.
We do this because the a in id doesn't mean the same as the a in (.). After that renaming we get:

(.) :: (b -> c) -> (a -> b) -> a -> c
id :: a1 -> a1
even :: Number -> Bool
0 :: Number

We change the as in id for a1. Okay, now that we won't have any collisions in type names, we GENERATE CONSTRAINTS:
We do this by putting each value's type against the expected type of that parameter in the composition:
(   b   ->   c ) -> (a  ->  b) ->   a    -> c

(Number -> Bool) -> (a1 -> a1) -> Number

This means that (Number -> Bool) should be equal to (b -> c), (a1 -> a1) to (a -> b), etc.

A constraint looks like a ~ b, where a and b are types, in this case, the list of constraints we originally get is:

(Number -> Bool) ~ (b -> c)
(a1 -> a1) ~ (a -> b)
a ~ Number

Function constraints can be splitted into simpler constraints as well, so we end with this final list of constraints:

Number ~ b
Bool ~ c
a1 ~ a
a1 ~ b
a ~ Number

We also keep a track of a dictionary from parametric types to other types that we will end up using to know what should type variable should mean, it starts empty: {}.

Now, we take the first constraint, Number ~ b. When we have a parametric type and a concrete type that does type check and what we do is delete that constraint, replace the parametric type (b) by the concrete type in all further appearences in other constraints, and add it to the dictionary:

Bool ~ c
a1 ~ a
a1 ~ Number <-- here we replaced b by Number
a ~ Number

{b: Number}

We continue one by one:

a1 ~ a
a1 ~ Number
a ~ Number

{b: Number, c: Bool}

------ Solved a1 ~ a

a ~ Number <-- replaced a1 by a
a ~ Number

{b: Number, c: Bool, a1: a}

------ Solved  a ~ Number

Number ~ Number <-- replaced a by Number

Now, we have a constraint of 2 concrete types. Since they both are the same type, it does type check! And we end up with:

{b: Number, c: Bool, a1: Number, a: Number} <-- we also replace it if the constraint appeared here

------ Solved Number ~ Number

We solved all constraints. That means the type checked!, and we can replace the parametric types in the composition type:

(b -> c) -> (a -> b) -> a -> c
Becomes:
(Number -> Bool) -> (Number -> Number) -> Number -> Bool

And since we both know that: the type checkes! + we passed the first 3 parameters, the type of
(even . id $ 0) is just the last parameter, Bool.

An example that doesn't type check:

(even . even)

(.) :: (b -> c) -> (a -> b) -> a -> c
even :: Number -> Bool

Constraints:

(Number -> Bool) ~ (b -> c)
(Number -> Bool) ~ (a -> b)

Expanded:

Number ~ b
Bool ~ c
Number ~ a
Bool ~ b

dictionary: {}

-------- Solved Number ~ b

Bool ~ c
Number ~ a
Bool ~ Number

{b: Number}

------ Solved Bool ~ c

Number ~ a
Bool ~ Number

{b: Number, c: Bool}

----- Solved Number ~ a

Bool ~ Number

{b: Number, c: Bool, a: Number}

Ooops, now we have a constraint that says Bool ~ Number, that's not true! So this type doesn't check.